backend/requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
pyrebase4>=4.7.1
google-cloud-storage>=2.14.0
setuptools>=80.9.0 